requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysql-connector-python-rf==2.2.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.0
